\name{bigJoin-package}
\alias{bigJoin-package}
\alias{bigJoin}
\docType{package}
\title{Enables efficient joining of data file on common fields using
the Unix utility join}
\description{wrapper around the Unix join facility which is much more efficient than the built-in R routine merge().
The package enables the joining of multiple files onn disk at once.
The files can be compressed and various filters can be deployed before joining.
Compiles only under Unix.}
\details{
\tabular{ll}{Package: \tab bigJoin\cr
Type: \tab Package\cr
Title: \tab Enables efficient joining of data file on common fields using
the Unix utility join\cr
Version: \tab 0.1\cr
Date: \tab 2015-02-15\cr
Depends: \tab R (>= 2.10), graphics, stats\cr
Author: \tab "Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>\cr
Maintainer: \tab "Markus Loecher" <markus.loecher@gmail.com>\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr}
}
\author{"Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>}

\keyword{ package }


