\name{JoinData-package}
\alias{JoinData-package}
\docType{package}
\title{Enables efficient joining of data file on common fields using the Unix utility join}
\description{This package is basically a convenience wrapper around the Unix join facility which is much more efficient than the built-in R routine merge().
The package enables the joining of multiple files at once.}
\details{
\tabular{ll}{Package: \tab JoinData\cr
Type: \tab Package\cr
Title: \tab Enables efficient joining of data file on common fields using the Unix utility join\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-01\cr
Depends: \tab \cr
Suggests: \tab \cr
Author: \tab "CodeAndMath" <markus@CodeAndMath.com>\cr
Maintainer: \tab "CodeAndMath" <markus@CodeAndMath.com>\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr}
}
\author{"CodeAndMath" <markus@CodeAndMath.com>}

\keyword{ package }


