\name{MultiJoin-package}
\alias{MultiJoin-package}
\alias{MultiJoin}
\docType{package}
\title{Enables Efficient Joining of Data File on Common Fields using the Unix Utility Join}
\description{Wrapper around the Unix join facility which is more efficient than the built-in R routine merge().
The package enables the joining of multiple files on disk at once.
The files can be compressed and various filters can be deployed before joining.
Compiles only under Unix.}
\details{
\tabular{ll}{Package: \tab MultiJoin\cr
Type: \tab Package\cr
Title: \tab Enables Efficient Joining of Data File on Common Fields using the Unix Utility Join\cr
Version: \tab 0.1\cr
Date: \tab 2015-05-07\cr
Depends: \tab R (>= 2.10), graphics, stats\cr
Author: \tab "Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>\cr
Maintainer: \tab "Markus Loecher" <markus.loecher@gmail.com>\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr}
}
\author{"Markus Loecher, Berlin School of Economics and Law (BSEL)" <markus.loecher@gmail.com>}

\keyword{ package }


